dagwriter: fix reuse of buffers problem.
The buffers from the client were getting reused (not copied)
and so the data got duplicated internally. The tests don't
catch this!! I tested manually by adding a large text (aeneid).
We should make a test for it.

And, FWIW, tried making the channel sync (i.e. dagwriter
splChan not be buffered), which may be the right way to fix
the problem, but didnt work.