t0080: prepare for online tests
Put the tests into a function so that they
can easily be run both offline and online.

Unfortunately it is not possible to run them
both offline and online right now because there
are some "Error: context deadline exceeded"
when the online tests are run if the offline
tests have been run before. This is explained
in a comment and the online tests are commented
out for now.

License: MIT
Signed-off-by: Christian Couder <chriscool@tuxfamily.org>