godep: vendor correctly.
I'm so sick of this absolute crap. It goes wrong all the time.
I want to get off godeps soon. But for now, let's **please**
setup a test on travis that exercises all sorts of godep things
to avoid merging anything that will break.

Maybe we can fail the build if any library is NOT vendored.