core/corehttp/gateway_test: Replace coreunix.Add with importer.BuildDagFromReader
coreunix.Add is basically a wrapper around BuildDagFromReader that
returns a string-ified key instead of a DAG node.  That's what we want
here, so it's a few more lines to inline the key extraction (mostly
due to Go's explicit error handling [1], which makes method chaining
not idomatic [2]).  But I think a bit of local error-checking is a
reasonable price for a more concise API, where there's only one
function for adding file nodes from a reader.

[1]: https://golang.org/doc/faq#exceptions
[2]: http://stackoverflow.com/a/27300387