Replace single callbacks with slices of function references
This allows you to easily pass in arrays of independent callbacks.
For example, you might check several conditions to decide whether or
not to ignore a file, and you can handle each condition in an
independent preNodeCallBack.  Or you might want to wrap your file
nodes in metadata nodes, and then print progress on those metadata
nodes to the terminal.  One tricky bit about the post-node callback
slice is that with the single-callback form the callback was called
with every newly-created node.  But with the new logic you can have:

1. Add creates node1.
2. First callback wraps node1 in node2 (so node2 links node1).
3. Second callback acts on node2, but is never called for node1.

However you could also have:

1. Add creates node1.
2. First callback drops node1 and replaced it with node2 (so node2
   does *not* depend on node1).
3. Second callback acts on node2, and is never called for node1, as it
   should be.

And you could also have:

1. Add creates node1.
2. First callback wraps node1 in node2 (so node2 links node1), and
   also creates a new node3 that is linked from node2:

     node2
     |-- node1
     `-- node3

3. Second callback acts on node2, but is never called for node1 or
   node3.

The only way I can think of to cleanly handle this case is to have the
postNodeCallback return a slice of new nodes, with the first node in
the slice being the new tip, and the subsequent nodes just being
passed on to the subsequent callbacks with top always unset.  That's
ugly enough that I'm putting it off for now ;).

We didn't have this problem with the previous single-callback
approach, but then you'd need a wrapping function or stateful object
to manage the union of callbacks in all cases.  With the current
implementation, folks can handle simple cases using the array of
callbacks, and fall back to creating a multi-callback wrapper when the
current iteration logic falls down for them.