core/commands/object: Intermediate-node creation for patch add-link
Because:

  ipfs object patch $ROOT add-link a/b/c $FILE

is a lot easier than:

  EMPTY=$(ipfs object new unixfs-dir) &&
  A=$(ipfs object patch $EMPTY add-link b $EMPTY) &&
  R=$(ipfs object patch $ROOT add-link a $A) &&
  ipfs object patch $R add-link a/b/c $FILE

and the long form isn't even checking to see if the original $ROOT has
descendents a or a/b.

Note that these are just Merkle nodes, not Unix-FS directories,
because 'ipfs object ...' is a Merkle-level tool.  I'd like to make
this flexible enough that we could have a Unix-FS-level 'ipfs file
patch ...' with similar semantics except that it operates on
Unix-FS-level nodes (e.g. it creates intermediate *directories*, turns
directories into files if you use 'set-data', manages the '*-data'
commands without clobbering the type information unixfs stores in
Data, etc.).  But I don't want that Unix-FS abstraction stuff sneaking
into the Merkle-level command we're working on here.

Now that insertNodeAtPath takes a *dag.Node to insert instead of a
key.Key, and it can both add and remove (if toinsert is nil) links,
there wasn't much need for the addLink helper.  This commit just
inlines (and extends to removal) that functionality in
insertNodeAtPath.

The '$(<file)' constructs differ from this file's previous '$(cat
file)' usage, but it's always good to remove useless cats ;).  I
adjusted the existing '$(cat file)' call while I was at it.

License: MIT
Signed-off-by: W. Trevor King <wking@tremily.us>