t0051: Add patch set-data tests
The "raw bytes as a command-line argument" UI makes it difficult to do
some things.  For example, I wasn't able to figure out a way to use

  $ ipfs object patch $EMPTY set-data "$(<set_data_expected)"

to add data from a file that had a trailing newline (which is why I
ended up using printf to write a file without a trailing newline).

License: MIT
Signed-off-by: W. Trevor King <wking@tremily.us>