Command handlers use whitelist expose commands
This PR adds the ability to expose commands to the gateway.

Each commands handler uses a whitelist to expose the commands
on the http interface.

All commands listed under the root command are exposed by default
to the commands port.

Only commands with the `GatewayAccess` flag will be exposed on the
gateway.

License: MIT
Signed-off-by: Travis Person <travis.person@gmail.com>