Treat symlinks as symlinks symlinks
On Linux, open() will follow symlinks unless you use the O_NOFOLLOW
flag (added in Linux 2.1.126).  Then it will error out with ELOOP
unless you've set O_PATH.  Unfortunately, Go doesn't define O_PATH and
has no plans to do so [1].

The other tricky bit here is extracting the errno [2] from the
*PathError [3] returned by os.Open [4].  The new files.Errno tries to
cast the error instance to an *os.PathError.  If that succeeds, it
looks at the *os.PathError's Err attribute and tries to cast that to
an Errno.  If that succeeds it returns the extracted errno.

With this commit, we now have File object for the symlinks with the
correct filename and mode.  I've also created a string-based reader to
pass along the link target, so the adder will have it available when
it gains support for these types of files.

[1]: https://github.com/golang/go/issues/7830
[2]: https://golang.org/pkg/syscall/#Errno
[3]: https://golang.org/pkg/os/#PathError
[4]: https://golang.org/pkg/os/#Open

License: MIT
Signed-off-by: W. Trevor King <wking@tremily.us>