Basic implementation of "add --no-copy".
This involved:

1) Constructing an alternative data object that instead of raw bytes is
a DataPtr with information on where the data is on the file system and
enough other information in AltData to reconstruct the Merkle-DAG node.

2) A new datastore "filestore" that stores just the information in
DataPtr.  When retrieving blocks the Merkle-DAG node is reconstructed
from combining AltData with the data from the file in the file system.
Because the datastore needs to reconstruct the node it needs access to
the Protocol Buffers for "merkledag" and "unixfs" and thus, for now,
lives in go-ipfs instead of go-datastore.

The filestore uses another datastore to store the protocol buffer
encoded DataPtr.  By default this is the leveldb datastore, as the size
fo the encoded DataPtr is small.

Towards #875.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>