Simpilfy files.File and chunk.Splitter interface.
Simplify files.File interface by combining Offset() and AbsPath()
methods into one that return a files.ExtraInfo interface that can be
extended with additional information.

Simplify chunk.Splitter by returning a Bytes struct in the NextBytes()
method.  This eliminates the need for the AbsPath() method and the
need to return the data and offset separately

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>