Basic implementation of "add --no-copy".
This involved:

1) Maintaing a DataPtr in merkledag.Node and blocks.Block.  The
DataPtr points to the location of the data within a file on the file
system.  It the node is a leaf it also contains an alternative
serialization of the Node or Block that does not contain the data.

3) A new datastore "filestore" that stores just the information in
DataPtr.  When retrieving blocks the Merkle-DAG node is reconstructed
from combining AltData with the data from the file in the file system.
Because the datastore needs to reconstruct the node it needs access to
the Protocol Buffers for "merkledag" and "unixfs" and thus, for now,
lives in go-ipfs instead of go-datastore.

The filestore uses another datastore to store the protocol buffer
encoded DataPtr.  By default this is the leveldb datastore, as the size
for the encoded DataPtr is small.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>