Filestore: Readd support for realtive directories with '-P' and '-l' options.
Add the '-P' (--physical) and '-l' (--logical) options to filestore
add to create absolute directories from relative ones.  The '-P'
options uses the physical (no symbolic links) location of the current
directory and the '-l' option uses the PWD env. variable if it is
available and if it names the current directory.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>