Filestore: Remove OrigData from ListRes and use iterators instead
The recent commit "Filestore: Fix a race condition when updating
DataObj" introduced a rather serious bug as the bytes returned through
a leveldb iterator Value() method could change on the next call to
Next() and I was using that result (without making a copy) in a
buffered channel.

Fix that bug by using Iterators instead when that field is required.
Iterators are more efficient than channels anyway.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>