Filestore: Fix maintenance commands so they are safe to run Online
And do so while avoiding the need for holding onto locks.

Maintenance commands are run on a snapshot of the database that is
guaranteed to be in a consistent state.  Basically a snapshot is taken
before a file is added.

To avoid a race condition when deleting a block that has since become
valid, the value of the snapshot is compared with the current value in
the database and if they don't match then the block is not deleted.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>