Filestore: Redo "clean" command so it works again and related changes.
The orignal "clean" command worked by doing up to three separate
passes and removing problem blocks after each pass.  The three-pass
approach depended on the blocks being deleted from the previous pass.
The change introduced in "Filestore: Fix maintenance commands so they
are safe to run Online" broke this by using a single snapshot for all
passes.  This commit reworks the "verify" command so that the
three-pass approach is no longer required.

Blocks are now normally only reported "incomplete" when one the child
nodes is missing, if a child node is just unreadable the status is now
reported as "problem".  The flag "incomplete-when" can change this
behavior.  This change eliminates the need for the second (incomplete)
pass.

To eliminate the need for the thired (orphan) pass in clean, a new
"verify-post-orphan" command is created.  This command works like
"verify" but in addition checks for orphans that would be created if
"incomplete" blocks are removed.

Also add more comprehensives tests of the "clean" command to the
sharness tests to make sure the "clean" does not break again.

Other various changes to the "verify" command including adding of a
"--no-obj-info" flag that only reports the status and hash.

Also fix special "append" status so that it is reported correctly.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>