Gateway: Replace Target Node Data when Exists on POST
Previously, if the specified path of a POST request
existed, the newly generated DAG would have a root
of the specified path instead of only modifying the
data of the existing DAG at that path. These changes
modify behavior so that the entire tree structure
of the initially specified root CID is maintained,
and the only change is the Data at the specified path.

License: MIT
Signed-off-by: Tom O'Donnell <todonnell@gmail.com>