Gateway PUT Overwrite: Fix for PUT on empty dir; added test
The previous commit would fail when an empty directory was
specified. This commit checks if the empty directory is
specified, and will return an empty directory as it did
previously. Also added a test case which verifies that a
PUT to an existing file for a given DAG will overwrite
its contents and return the correct new DAG.

License: MIT
Signed-off-by: Tom O'Donnell <todonnel91@gmail.com>