GC now has concurrent mark phase
Introduce GC rework with concurrent mark phase, tricolor set and rework
that in future will allow to incremental and/or concurrent mark and
sweep phases.

Currently it is made a bit difficult by direct pins.

It will also allow for reuse of colorset for improved performance and no
set reallocation.

From observations pre-rewrok it should reduce GC lock time by half as
about 50% of the time is spent in mark phase.
This is consistent with observations made after the implementation.

License: MIT
Signed-off-by: Jakub Sztandera <kubuxu@protonmail.ch>