core: get cmd outPath removes trailing slash - fixes #3729
When using path with trailing slashes, ipfs get command had issues
with trying to store content into '.' folder (as the file part of the path
was empty). Now, it correctly stores the file into desired folder,
thanks to the removal of trailing slash as for the get command arg.

Commit a09d97465270ef441791b65bca98d37c25eb0564 was a poor fix, which only
modified the path printed to the StdOut. The arguments of command are
passed to the request's reader, which is turned into a tar Extractor.
The headers of that the tar Extractor were incorrectly populating the
write path, missing the root directory, causing files nested in root
folder to be written into the root output path.

fix visible at: https://asciinema.org/a/mDili4YFnOacgt3dbjHuXTzcT

License: MIT
Signed-off-by: adamliesko <adamliesko@gmail.com>