get rid of race in WaitPub
Before, WaitPub could notice that there is something to publish, sleep, then
wake up and write to pubnowch *after* the value is published. This would cause
it to hang until the next publish.

Instead, *always* write to pubnowch and check to see if we *really* need to
publish inside the `publish` function.

License: MIT
Signed-off-by: Steven Allen <steven@stebalien.com>