fix logs
i changed the logs to be log.Error so they surface to stdout during the
boot process of the daemon.  --enable-mplex-experiment is set to true by
default, so when no config is specified, both yamex and mplex will be
used.  i would imagine that as flags like --enable-mplex-experiment
get phased out, the behavior will be more closely related to what is in
the config file.

License: MIT
Signed-off-by: Jack Loughran <mail@jloughran.com>