Restore ability to regenerate merkledag.pb.go
Due to the IPLD-PB format relying on a mis-feature of gogo-protobuf, the
process of (re)generating the golang protobuf bindings is way harder than
it ought to be :(

What I do here is compile-and-execute a patched version of the gogofast
encoder, and present it as a temporary generator-plugin to the stock `protoc`
binary. This ensures we still get the effect we want, but the "nonstandard"
generator does not get installed anywhere.

All tests verified to pass despite the seemingly substantial changes in the
generated code

License: MIT
Signed-off-by: Mib Kd743naq <mib.kd743naq@gmail.com>