importer: fillNodeRec: pass `UnixfsNode` structure, not its pointer
Modify `fillNodeRec` of the balanced importer to pass the `UnixfsNode` structure
instead of its pointer. The pointer was used as a channel to return information
from the recursive callee function which was obscuring the code information
flow. This returned information is now explicitly added to the return values of
`fillNodeRec` that now include the IPLD node to be added and its (UnixFS) file
size (that depends on the type of node and is handled by the encapsulating
`UnixfsNode` structure).

As a consequence of this, IPLD nodes can now be used instead of the composite
`UnixfsNode` in some parts of the code (which helps clarifying what type of node
is being manipulated in each case). Some of the functions that used `UnixfsNode`
as arguments were substituted with analogous functions that receive `ipld.Node`,
the original functions cannot be replaced (at the moment) because they are still
being used by the trickle builder.

The indirect modification of the passed `UnixfsNode` through its pointer was
used in the corner case of the `Layout` function when the loop is iterated only
once, i.e., the generated DAG is a single node with data and `AddChildNode` is
never called. In that case the `root` was converted through its pointer to a
leaf node with data in `fillNodeRec` which could later be added to the DAG
(`db.Add()`). This indirect modification is not allowed now so that corner case
needs to be handled separately by explicitly adding the `filledRoot` node
returned by `fillNodeRec` which wasn't added to `root` (instead of adding `root`
itself).

The priority of this commit is to respect the structure and code flow of the
importer to still make evident how a balanced DAG is built.

License: MIT
Signed-off-by: Lucas Molas <schomatis@gmail.com>